# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.APICreateCommandRequest:
    properties:
      args:
        additionalProperties:
          type: string
        type: object
      command:
        enum:
        - upgrade
        - upgrade-recovery
        - reset
        - apply-cloud-config
        - reboot
        - exec
        example: upgrade
        type: string
    type: object
...
  store.NodeCommand:
    properties:
      args:
        additionalProperties:
          type: string
        type: object
      command:
        type: string
      completedAt:
        type: string
      createdAt:
        type: string
      deliveredAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      managedNodeID:
        type: string
      phase:
        type: string
      result:
        type: string
    type: object
...
info:
  contact:
    name: Kairos authors
    url: https://github.com/kairos-io/AuroraBoot
  description: |-
    HTTP API for the AuroraBoot Kairos node manager.

    Three security schemes are used across the API:
    * AdminBearer — Authorization: Bearer <admin-password>. Human/UI/automation callers.
    * NodeAPIKey — Authorization: Bearer <api-key>. Registered Kairos nodes calling their own endpoints. The api-key is returned from POST /api/v1/nodes/register.
    * RegistrationToken — shared secret carried inside the body of POST /api/v1/nodes/register.

    Two WebSocket endpoints also exist but are not described in this spec:
    * GET /api/v1/ws — agent command channel (auth via ?token=<api-key>).
    * GET /api/v1/ws/ui — UI live update channel (auth via ?token=<admin-password>).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: AuroraBoot API
  version: 0.1.0
paths:
...
  /api/v1/nodes/{nodeID}/commands:
    post:
      consumes:
      - application/json
      parameters:
      - description: Node ID
        in: path
        name: nodeID
        required: true
        type: string
      - description: Command payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.APICreateCommandRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.NodeCommand'
      security:
      - AdminBearer: []
      summary: Queue a command for a single node
      tags:
      - Commands
...
securityDefinitions:
  AdminBearer:
    description: Supply as "Bearer <admin-password>".
    in: header
    name: Authorization
    type: apiKey
  NodeAPIKey:
    description: Supply as "Bearer <api-key>" returned from POST /api/v1/nodes/register.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"